fix: changes asked by reviewer
